+++ providers/src/airflow/providers/docker/decorators/docker.py
@@ -22,13 +22,13 @@ import warnings
 from tempfile import TemporaryDirectory
 from typing import TYPE_CHECKING, Any, Callable, Literal, Sequence
 
+from airflow.decorators.base import DecoratedOperator, task_decorator_factory
 from airflow.exceptions import AirflowException, AirflowProviderDeprecationWarning
 from airflow.providers.common.compat.standard.utils import write_python_script
 from airflow.providers.docker.operators.docker import DockerOperator
-from airflow.providers.standard.decorators.base import DecoratedOperator, task_decorator_factory
 
 if TYPE_CHECKING:
-    from airflow.providers.standard.decorators.base import TaskDecorator
+    from airflow.decorators.base import TaskDecorator
     from airflow.utils.context import Context
 
     Serializer = Literal["pickle", "dill", "cloudpickle"]
