+++ providers/src/airflow/providers/snowflake/decorators/snowpark.py
@@ -19,12 +19,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable, Sequence
 
+from airflow.decorators.base import DecoratedOperator, task_decorator_factory
 from airflow.providers.snowflake.operators.snowpark import SnowparkOperator
 from airflow.providers.snowflake.utils.snowpark import inject_session_into_op_kwargs
-from airflow.providers.standard.decorators.base import DecoratedOperator, task_decorator_factory
 
 if TYPE_CHECKING:
-    from airflow.providers.standard.decorators.base import TaskDecorator
+    from airflow.decorators.base import TaskDecorator
 
 
 class _SnowparkDecoratedOperator(DecoratedOperator, SnowparkOperator):
