+++ providers/src/airflow/providers/standard/decorators/branch_python.py
@@ -18,12 +18,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable
 
-from airflow.providers.standard.decorators.base import task_decorator_factory
+from airflow.decorators.base import task_decorator_factory
 from airflow.providers.standard.decorators.python import _PythonDecoratedOperator
 from airflow.providers.standard.operators.python import BranchPythonOperator
 
 if TYPE_CHECKING:
-    from airflow.providers.standard.decorators.base import TaskDecorator
+    from airflow.decorators.base import TaskDecorator
 
 
 class _BranchPythonDecoratedOperator(_PythonDecoratedOperator, BranchPythonOperator):
