+++ providers/src/airflow/providers/standard/decorators/sensor.py
@@ -19,11 +19,11 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable, ClassVar, Sequence
 
-from airflow.providers.standard.decorators.base import get_unique_task_id, task_decorator_factory
+from airflow.decorators.base import get_unique_task_id, task_decorator_factory
 from airflow.providers.standard.sensors.python import PythonSensor
 
 if TYPE_CHECKING:
-    from airflow.providers.standard.decorators.base import TaskDecorator
+    from airflow.decorators.base import TaskDecorator
 
 
 class DecoratedSensorOperator(PythonSensor):
